:package: xcode group reference for addons
for the ofxKinect addon.  If the addon wasn't already referenced from an example template, the best community recommendation is to drag the files into the addons group in the xcode project view.  This gets the files added to the project includes paths.  It's not the most well documented part of openFrameworks.